pyproject.toml
[project]
name = "negative-price-calc"
version = "0.1.1"
description = "Python application for analyzing electricity prices and solar production data, focusing on negative price detection and cost analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "chardet>=5.0.0",
    "entsoe-py>=0.6.9",
    "openpyxl>=3.1.2",
]

[project.scripts]
negative-price-calc = "main:main"
se-cli = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "utils", "cli"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
